refactored API options to make them testable
