+++ library/Data/Numbers/Trivia/Api.hs
@@ -21,7 +21,9 @@ module Data.Numbers.Trivia.Api (
 import Network.HTTP.Simple
 import Data.Function ((&))
 import Data.List (intercalate)
-import Data.ByteString.Char8 (pack, unpack, ByteString)
+import Data.ByteString.Char8 (unpack, ByteString)
+
+import Data.Numbers.Trivia.Api.Options
 
 apiHost :: ByteString
 apiHost = "numbersapi.com"
@@ -57,22 +59,6 @@ data ApiEndpoint = GetTrivia Number
                  | GetDate Date
                  | GetRandom
 
--- | API options that can be used with any '…With' method
-type ApiOptions = [ApiOption]
-
--- | Possible API options that can be used on most endpoints
-data ApiOption = Fragment -- ^ returns the fact as a sentence fragment
-               | DefaultMsg ByteString -- ^ sets the default response message if requested number has no entry
-               | NotFound NotFoundAction -- ^ selects an alternative action if requested number has no entry
-               | MinLimit Int -- ^ lower limit (inclusive) when requesting random values
-               | MaxLimit Int -- ^ upper limit (inclusive) when requesting random values
-               deriving Show
-
--- | Describes an alternative action if a requested number has no entry
-data NotFoundAction = CeilIfNotFound -- ^ use the next higher number which has an entry
-                    | FloorIfNotFound -- ^ use the next lower number which has an entry
-                    | DefaultIfNotFound -- ^ return the default response
-
 instance Show Number where
   show (Number x) = show x
   show (Range xs) = show <$> xs & intercalate ","
@@ -92,31 +78,6 @@ instance Show ApiEndpoint where
   show (GetYear   x) = show x ++ "/year"
   show (GetDate   d) = show d ++ "/date"
   show  GetRandom    = "random"
-
-instance Show NotFoundAction where
-  show CeilIfNotFound = "ceil"
-  show FloorIfNotFound = "floor"
-  show DefaultIfNotFound = "default"
-
--- | By default, no API options are used
-defaultApiOptions :: ApiOptions
-defaultApiOptions = []
-
--- | Helper function to show a value as BS.ByteString
-bshow :: Show a => a -> ByteString
-bshow = pack . show
-
--- | Builds a key-value-pair from an API option
-optionToQueryItem :: ApiOption -> QueryItem
-optionToQueryItem  Fragment      = ("fragment", Nothing)
-optionToQueryItem (DefaultMsg m) = ("default", Just m)
-optionToQueryItem (NotFound   a) = ("notfound", Just $ bshow a)
-optionToQueryItem (MinLimit   x) = ("min", Just $ bshow x)
-optionToQueryItem (MaxLimit   x) = ("max", Just $ bshow x)
-
--- | Builds a query string from given API options
-optionsToQuery :: ApiOptions -> Query
-optionsToQuery = map optionToQueryItem
 
 -- | Builds an HTTP request with given API options and endpoint description
 makeApiRequestWith :: ApiOptions -> ApiEndpoint -> Request
